web/api/requirements.txt
fastapi==0.111.0
python-multipart
python-dotenv
ldap3
pyjwt
jose
httpx
cryptography
uvicorn[standard]==0.30.1
sqlmodel==0.0.16
aiosmtplib==2.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
jinja2==3.1.4
requests==2.32.2
libvirt-python==10.0.0
